pre-christmas checkin

+++ source/matlab/learn_manhattan_dp.m
@@ -1,22 +1,27 @@
-function [weights cases]=learn_manhattan_dp(cases)
+function [weights loaded_cases]=learn_manhattan_dp(cases)
 % learn_manhattan_dp
 %   Entry point to discriminative training for manhattan DP reconstructor
   
   % pick an output directory
   v = clock;
   basedir = '/home/alex/Code/indoor_context/source/matlab';
-  outdir = sprintf('%s/runs/%d:%d.%d %d-%d-%d', basedir, round(v([4 5 6 3 2 1])));
+  outdir = sprintf('%s/runs/%02d:%02d.%02d %02d-%02d-%04d', basedir, round(v([4 5 6 3 2 1])));
   r = mkdir(outdir);
   check r;
   
   % create symbolic links to this dir
-  linkpath = [outdir '/lastrun'];
-  system(['rm ' linkpath]);
-  system(['ln -s ' outdir ' ' linkpath]);
+  linkpath = [basedir '/lastrun'];
+  if exist(linkpath)
+    system(['rm ' linkpath]);
+  end
+  cmd = ['ln -s ''' outdir ''' ''' linkpath ''''];
+  report cmd;
+  system(cmd);
 
   % load cases
   if nargin==0
     cases = dp_load_cases('lab_kitchen1', 5:5:90, 'default');
+    loaded_cases = cases;
   end
 
   randn('state', 0);
@@ -52,7 +57,22 @@ function [weights cases]=learn_manhattan_dp(cases)
 
   margin_calls = 0;
   
-  model = svm_struct_learn(' -c 1.0 -o 2 -v 1 ', sparm) ;
+  % -v: verbosity for structSVM
+  % -y: verbosity for svm_light
+  % -b: percentage of training examples to renew examples for (as
+  %     percentage in [0,100])
+  % -f: number of constraints to cache
+  % -o: type of rescaling (margin or slack)
+  common_args = ' -o 2 -v 3 -y 1 ';
+  basic_args = [common_args ' -c 0.00001'];
+  caching_args = [common_args ' -c 0.01  -w 4  -f 5 '];
+  svm_struct_args = caching_args;
+
+  begin_learn_time = tic;
+  last_eval_time = tic;
+  last_accuracy = -1;
+  
+  model = svm_struct_learn(svm_struct_args, sparm) ;
 
   % ------------------------------------------------------------------
   %                                                         Evaluation
@@ -67,7 +87,7 @@ function [weights cases]=learn_manhattan_dp(cases)
   function delta = lossCB(param, estimated, ground_truth)
 	  disp('### Computing loss ###');
     delta = get_pixel_loss(estimated, ground_truth);
-    check prod(size(delta)) == 1;
+    %check prod(size(delta)) == 1;
     disp('# Done. #')
   end
 
@@ -89,17 +109,31 @@ function [weights cases]=learn_manhattan_dp(cases)
   % find margin-rescaling largest violation
   % argmax_y delta(yi, y) + < psi(x,y), w >
 	  disp('### Finding most-violated constraint ###');
+    disp(['                                      Last Accuracy: ' ...
+          num2str(last_accuracy) '%']);
 
     weights = unpack_weights(model.w', pix_ftr_size, wall_ftr_size);
 
     margin_calls = margin_calls + 1;
-    if mod(margin_calls, length(cases)*5) == 0
-      outfile = sprintf('%s/iter%04d.mat', outdir, margin_calls);
-      save(outfile, 'weights');
+    iter = floor(margin_calls/length(cases));
+    elapsed_time = toc(begin_learn_time);
+    if (mod(margin_calls, 10) == 0)
+      % report time
+      disp(' ');
+      disp(['Averaging ' num2str(1000*elapsed_time/margin_calls) ...
+            'ms per call to marginCB']);
+      disp(' ');
+    end
+    
+    time_since_eval = toc(last_eval_time);
+    %if mod(margin_calls, length(cases)*5) == 0
+    if time_since_eval > 60   % evaluate solution once per minute
+      outfile = sprintf('%s/iter%04d.mat', outdir, iter);
+      save(outfile, 'weights', 'svm_struct_args', 'iter');
       
       disp(['Called ' num2str(margin_calls) ' times so far']);
-      disp(['Called ' num2str(margin_calls) ' times so far']);
-      evaluate(cases, weights);
+      last_accuracy = evaluate(cases, weights);
+      last_eval_time = tic;
     end
     
     obj = make_contra_objective(casedata.pixel_features, ...
